Create README.md
Hi, I'm allenm737 and I understand that you may be able to help me with me with drops on Twitch. In the not so distant past, I was receiving drops on a regular basis from the CCs or reviewer's. occasionally, I may receive items from drops but not like I have in the past, very slim! I didn't want to make this to long and boring, so that is what's going on. Thank you so much for your time in advance.